refactor: extract config tests and split modules
