+++ src/chat/mod.rs
@@ -1,9 +1,9 @@
 // Core service types
 pub use self::service::{ChatService, ChatState};
-pub use self::service::{subscribe, unsubscribe, send, status, emotes, events};
+pub use self::service::{emotes, events, send, status, subscribe, unsubscribe};
 
 // Internal modules - keep private
-mod events;
 mod emotes;
+mod events;
 mod irc;
 mod service;
